+++ src/raft/state/watcher.rs
@@ -7,7 +7,7 @@ use tokio::sync::oneshot;
 
 use super::Result;
 
-type LockedUintMap<V> = Mutex<HashMap<u128, V>>;
+type LockedUintMap<V> = Mutex<HashMap<u64, V>>;
 type Bytes = Vec<u8>;
 
 #[derive(Debug, Default)]
@@ -18,13 +18,13 @@ pub struct Watcher {
 }
 
 impl Watcher {
-    pub fn register_command_watch(&self, idx: u128) -> oneshot::Receiver<Result<Bytes>> {
+    pub fn register_command_watch(&self, idx: u64) -> oneshot::Receiver<Result<Bytes>> {
         let (tx, rx) = oneshot::channel();
         self.command_watches.lock().unwrap().insert(idx, tx);
         rx
     }
 
-    pub fn command_applied(&self, idx: u128, result: Result<Bytes>) {
+    pub fn command_applied(&self, idx: u64, result: Result<Bytes>) {
         match self.command_watches.lock().unwrap().remove(&idx) {
             Some(submit_tx) => {
                 // If we have an error here its because the receiver hung up.
@@ -35,13 +35,13 @@ impl Watcher {
         }
     }
 
-    pub fn register_registration_watch(&self, idx: u128) -> oneshot::Receiver<()> {
+    pub fn register_registration_watch(&self, idx: u64) -> oneshot::Receiver<()> {
         let (tx, rx) = oneshot::channel();
         self.registration_watches.lock().unwrap().insert(idx, tx);
         rx
     }
 
-    pub fn registration_applied(&self, idx: u128) {
+    pub fn registration_applied(&self, idx: u64) {
         match self.registration_watches.lock().unwrap().remove(&idx) {
             Some(submit_tx) => {
                 let _ = submit_tx.send(());
@@ -50,13 +50,13 @@ impl Watcher {
         }
     }
 
-    pub fn register_cluster_config_watch(&self, idx: u128) -> oneshot::Receiver<()> {
+    pub fn register_cluster_config_watch(&self, idx: u64) -> oneshot::Receiver<()> {
         let (tx, rx) = oneshot::channel();
         self.cluster_config_watches.lock().unwrap().insert(idx, tx);
         rx
     }
 
-    pub fn cluster_config_applied(&self, idx: u128) {
+    pub fn cluster_config_applied(&self, idx: u64) {
         match self.cluster_config_watches.lock().unwrap().remove(&idx) {
             Some(submit_tx) => {
                 let _ = submit_tx.send(());
